requirements.txt
torch>=2.6.0
fbgemm_gpu>=1.1.0; platform_system == "Linux"
torchrec>=1.1.0; platform_system == "Linux"
torchrec<1.1.0; platform_system != "Linux"
gin_config>=0.5.0
pandas>=2.2.0
tensorboard>=2.19.0
pybind11
